fix polyfills using expo-dev-client and platform splitting
