+++ react-native/.gitignore
@@ -9,6 +9,8 @@ npm-debug.*
 *.mobileprovision
 *.orig.*
 web-build/
+android
+ios
 
 # macOS
 .DS_Store
