+++ react-native/package.json
@@ -9,13 +9,16 @@
     "test": "jest",
     "lint": "expo lint",
     "generate": "buf generate buf.build/connectrpc/eliza",
-    "ci": "npm run generate && npm run build && npm run test"
+    "ci": "npm run generate && npm run build && npm run test",
+    "android": "expo run:android",
+    "ios": "expo run:ios"
   },
   "dependencies": {
     "@expo/vector-icons": "^14.0.2",
     "@react-navigation/native": "^6.1.17",
     "expo": "^51.0.17",
     "expo-constants": "^16.0.2",
+    "expo-dev-client": "^4.0.19",
     "expo-font": "^12.0.7",
     "expo-linking": "^6.3.1",
     "expo-router": "^3.5.17",
